Use a back-to-front buffer to optimize encoding
In order to encode a TNetstring sequentially, the length of the encoded
string must first be known. This requires recursive encoding of all
elements. Instead, the string can be encoded from the end, starting with
the type indicator, the body and then the length. This allows encoding
to be progressive and gives a significant (~2.5x) performance increase.

Encoding before: 32461/s
Encoding after:  84581/s

The idea came from Ryan Kelly on the mongrel2 mailing list[1],
suggesting to render tnetstrings in reverse.

[1] http://librelist.com/browser//mongrel2/2011/3/21/tnetstrings-speed-te
	st/#6c1c374f4add2c479fe4dd81594584dc