Use SvCUR() instead of strlen() to get string length
The benefit of having the size of the payload in the prefix is that NULL
byte are allowed in the payload. By using strlen() NULL bytes
effectively truncate the string. Use SvCUR to get the real string
length, and subsequently use the size prefix to determine payload
boundries.

Since the (sub)string is not traversed on each tn_decode() performance
has also improved by about 10% for decoding.