Use the `z` length modifier when formatting `STRLEN`
`STRLEN` is a `size_t`, which has its own length modifier since C99 in
`sprintf`-like format strings. However `%d` is being used which has
a different size. Use the `z` modifier for STRLEN.

Note that this change is not portable for older compilers/platforms
where the `z` modifier is not implemented.